openssl: Use the c_rehash shell re-implementation for target
We had a c_rehash shell re-implementation being used for the native
package and there is no reason to not use it as well for the
target. This allows it to be available without the need of perl being
installed.

This partially reverts OE-Core:d2b1a889ef (openssl: move c_rehash pkg
to avoid perl dep) but still allows the removal of perl
dependency.

The respective re-implementation was already in use here at OE-Core by
the OpenSSL 1.0 recipe (since 2016) and were removed from the target
recipe during the upgrade to the 1.1 release. So it now aligns the
native and target recipes.

,----
| Author: Otavio Salvador <otavio@ossystems.com.br>
| Date:   Mon May 23 17:45:25 2016 -0300
|
|     openssl: Add Shell-Script based c_rehash utility
|
|     The PLD Linux distribution has ported the c_rehash[1] utility from
|     Perl to Shell-Script, allowing it to be shipped by default.
|
|     1. https://git.pld-linux.org/?p=packages/openssl.git;a=blob;f=openssl-c_rehash.sh;h=0ea22637ee6dbce845a9e2caf62540aaaf5d0761
|
|     The OpenSSL upstream intends[2] to convert the utility for C
|     however did not yet finished the conversion.
|
|     2. https://rt.openssl.org/Ticket/Display.html?id=2324
|
|     This patch adds this script and thus removed the Perl requirement
|     for it.
|
|     Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>
|     Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>
`----

Series-to: oe-core
Series-version: 2
Series-changes: 2
- updated commit log

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>